Change how uffd is configured with Wasmtime
This commit updates how the `uffd` feature of Wasmtime is enabled and
used. Previously to use `uffd` users needed to do a few things:

* Enable the pooling allocator
* Enable paged memory initialization
* Enable the `uffd` feature at compile time

If any one of these were omitted, however, then `uffd` was silently not
used or otherwise wasn't beneficial (e.g. forgetting paged memory
initialization means that memories would be all initialized
immediately).

This commit changes this configuration to instead have consumers
explicitly request `uffd` as part of enabling the pooling allocation
strategy. This makes it clear that `uffd` is only enabled for the
pooling allocator, and requests to enable it when the feature isn't
compiled in (or is the wrong platform) can return an error indicating
what happened.

This does sort of effectively revert #3245 which allows enabling paged
memory initialization independent of uffd, but at this time I can't
quite remember what the motivation was for that PR so I'm hoping that
this can jog a memory!