Pooling allocator: make backend (mmap, uffd, memfd) runtime-configurable.
Currently, the pooling allocator makes heavy use of compile-time
features to determine its runtime behavior and allocation strategy. This
makes testing harder, and makes runtime configuration from higher levels
of the system impossible.

This PR passes a `PoolingBackend` value down from the configuration
plumbing and into the pooling allocator, and uses its value to
dynamically choose a backend instead. This is mostly a matter of lots of
plumbing and conditionals.

The most complex part of this PR is actually the testing and feature
setup. We still want the memfd and uffd backends to be optional -- i.e.,
one should be able to build a wasmtime binary without them included.
This implies not including the enum options in the config enum, and
everything down from there.

In order to test all possible situations, we actually need to test four
builds -- with neither uffd nor memfd, with just one, or with both. We
also want to test that everything works when different backends are
specified, and the simplest way to do this is just to run the whole test
suite with the default backend set to each possible option. Hence we run
tests with both backends included and the default set three different
ways; with one backend included and the default set to that or the
fallback mmap backend; etc.

I'm not sure that this is actually the simplest option -- it may, in the
end, be simpler to entirely remove uffd, if we eventually determine that
memfd supersedes it, and then unconditionally include memfd support on
Linux. But it was suggested here [1] to opt for a dynamic approach, so
this PR explores what that would look like!

[1]
https://github.com/bytecodealliance/wasmtime/pull/3697#discussion_r792246996